# Reuse rune array instead of computing it every time.