# PR Details

Reuse rune array instead of computing it every time.

## Description

Converting a string to a rune array is not cheap. Since Golang strings are immutable, each call to currentChar, doubleChar, etc. is doing a lot of unnecessary work.

This commit fixes this by storing the conversion of string to runes inside the parser and using it instead. This still incurs in the cost of conversion, but it's done only once.

## Related Issue

https://github.com/xuri/nfp/issues/3

## Motivation and Context

We noticed that parsing the file raw was a lot faster than parsing it when parsing numbers. I looked into this library and the rune conversion immediately jumped out as a bottleneck.

## How Has This Been Tested

I added a benchmark to compare this change against the main branch:

With my changes:

```
~/w/nfp ❯❯❯ go test -v -bench=. -benchmem                                                                        main ⬆ ✱
=== RUN   TestNFP
--- PASS: TestNFP (0.01s)
goos: darwin
goarch: amd64
pkg: github.com/xuri/nfp
cpu: Intel(R) Core(TM) i7-9750H CPU @ 2.60GHz
BenchmarkNumberFormatParser
BenchmarkNumberFormatParser-12    	     548	   2141649 ns/op	  784948 B/op	   25909 allocs/op
PASS
ok  	github.com/xuri/nfp	1.440s
```

Before my changes:
```
=== RUN   TestNFP
--- PASS: TestNFP (0.02s)
goos: darwin
goarch: amd64
pkg: github.com/xuri/nfp
cpu: Intel(R) Core(TM) i7-9750H CPU @ 2.60GHz
BenchmarkNumberFormatParser
BenchmarkNumberFormatParser-12    	     127	   9484347 ns/op	 8309766 B/op	   65128 allocs/op
PASS
ok  	github.com/xuri/nfp	2.204s
```

## Types of changes

- [ ] Docs change / refactoring / dependency upgrade
- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)

## Checklist

- [x] My code follows the code style of this project.
- [ ] My change requires a change to the documentation.
- [ ] I have updated the documentation accordingly.
- [x] I have read the **CONTRIBUTING** document.
- [x] I have added tests to cover my changes.
- [x] All new and existing tests passed.